+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from "react"
 import ReactDOM from "react-dom"
-import App from "./components/App"
+import App from "./App"
 import "bootstrap/dist/css/bootstrap.min.css"
 
 ReactDOM.render(
